requirements.txt
psycopg2-binary>=2.9.9
pandas>=2.1.0
numpy>=1.24.0
faker>=20.0.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
jupyter>=1.0.0